New attempt to get song playing working
I managed to get it to work using multithreading, and it works ok, but
the timing isn't perfect and sometimes there are slight audio glitches.
I'm thinking maybe what I should do is just mix all of the sounds down
to a single song and just play that. I'll work a little more on
multithreading first though.
Other changes:
* Made all of the samples mono
* Removed some tags from the samples as they were causing clicks to be
played